+++ packages/dht/src/helpers/createPeerDescriptor.ts
@@ -1,4 +1,4 @@
-import { ECDSA_SECP256K1_EVM } from '@streamr/utils'
+import { EcdsaSecp256k1Evm } from '@streamr/utils'
 import crypto from 'crypto'
 import { isBrowserEnvironment } from '../helpers/browser/isBrowserEnvironment'
 import { createPeerDescriptorSignaturePayload } from '../helpers/createPeerDescriptorSignaturePayload'
@@ -9,7 +9,9 @@ import {
     PeerDescriptor
 } from '../../generated/packages/dht/protos/DhtRpc'
 
-const calculateNodeIdRaw = (ipAddress: number, privateKey: Uint8Array): DhtAddressRaw => {
+const signingUtil = new EcdsaSecp256k1Evm()
+
+const calculateNodeIdRaw = async (ipAddress: number, privateKey: Uint8Array): Promise<DhtAddressRaw> => {
     // nodeId is calculated as 
     // concatenate(
     //   get104leastSignificatBits(hash(ipAddress)), 
@@ -17,8 +19,8 @@ const calculateNodeIdRaw = (ipAddress: number, privateKey: Uint8Array): DhtAddre
     // )
     const ipAsBuffer = Buffer.alloc(4)
     ipAsBuffer.writeUInt32BE(ipAddress)
-    const ipHash = ECDSA_SECP256K1_EVM.keccakHash(ipAsBuffer)
-    const signature = ECDSA_SECP256K1_EVM.createSignature(ipAsBuffer, privateKey)
+    const ipHash = signingUtil.keccakHash(ipAsBuffer)
+    const signature = await signingUtil.createSignature(ipAsBuffer, privateKey)
     const nodeIdRaw = Buffer.concat([
         ipHash.subarray(ipHash.length - 13, ipHash.length),
         signature.subarray(signature.length - 7, signature.length)
@@ -26,14 +28,15 @@ const calculateNodeIdRaw = (ipAddress: number, privateKey: Uint8Array): DhtAddre
     return nodeIdRaw
 }
 
-export const createPeerDescriptor = (connectivityResponse: ConnectivityResponse, region: number, nodeId?: DhtAddress): PeerDescriptor => {
+export const createPeerDescriptor = async (connectivityResponse: ConnectivityResponse, 
+    region: number, nodeId?: DhtAddress): Promise<PeerDescriptor> => {
     const privateKey = crypto.randomBytes(32)
     const publicKey = crypto.randomBytes(20)  // TODO calculate publicKey from privateKey
     let nodeIdRaw: DhtAddressRaw
     if (nodeId !== undefined) {
         nodeIdRaw = toDhtAddressRaw(nodeId)
     } else {
-        nodeIdRaw = calculateNodeIdRaw(connectivityResponse.ipAddress, privateKey)
+        nodeIdRaw = await calculateNodeIdRaw(connectivityResponse.ipAddress, privateKey)
     }
     const ret: PeerDescriptor = {
         nodeId: nodeIdRaw,
@@ -49,6 +52,6 @@ export const createPeerDescriptor = (connectivityResponse: ConnectivityResponse,
             tls: connectivityResponse.websocket.tls
         }
     }
-    ret.signature = ECDSA_SECP256K1_EVM.createSignature(createPeerDescriptorSignaturePayload(ret), privateKey)
+    ret.signature = await signingUtil.createSignature(createPeerDescriptorSignaturePayload(ret), privateKey)
     return ret
 }