+++ packages/sdk/src/identity/MLDSAKeyPairIdentity.ts
@@ -1,19 +1,17 @@
-import { hexToBinary, ML_DSA_87 } from '@streamr/utils'
+import { hexToBinary, SigningUtil } from '@streamr/utils'
 import { KeyPairIdentity } from './KeyPairIdentity'
 import { SignatureType } from '@streamr/trackerless-network'
 import { StrictStreamrClientConfig } from '../Config'
+
+const signingUtil = SigningUtil.getInstance('ML_DSA_87')
 
 /**
  * An identity that uses a quantum-resistant ML-DSA-87 key pair to sign messages.
  */
 export class MLDSAKeyPairIdentity extends KeyPairIdentity {
-    assertKeyPairIsValid(): void {
-        // Validity of key pair is tested by signing and validating something
-        const payload = Buffer.from('data-to-sign')
-        const signature = ML_DSA_87.createSignature(payload, this.privateKey)
-        if (!ML_DSA_87.verifySignature(this.publicKey, payload, signature)) {
-            throw new Error(`The given publicKey and privateKey don't match!`)
-        }
+
+    assertValidKeyPair(): void {
+        signingUtil.assertValidKeyPair(this.publicKey, this.privateKey)
     }
 
     // eslint-disable-next-line class-methods-use-this
@@ -21,20 +19,11 @@ export class MLDSAKeyPairIdentity extends KeyPairIdentity {
         return SignatureType.ML_DSA_87
     }
 
-    // eslint-disable-next-line class-methods-use-this
-    getExpectedPublicKeyLength(): number {
-        return 2592
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    getExpectedPrivateKeyLength(): number {
-        return 4896
-    }
-
     async createMessageSignature(payload: Uint8Array): Promise<Uint8Array> {
-        return ML_DSA_87.createSignature(payload, this.privateKey)
+        return signingUtil.createSignature(payload, this.privateKey)
     }
 
+    /** @internal */
     static fromConfig(config: Pick<StrictStreamrClientConfig, 'auth'>): MLDSAKeyPairIdentity {
         const keyPairConfig = KeyPairIdentity.getKeyPairFromConfig(config)
         if (!keyPairConfig.publicKey) {
@@ -44,7 +33,7 @@ export class MLDSAKeyPairIdentity extends KeyPairIdentity {
     }
 
     static generate(): MLDSAKeyPairIdentity {
-        const keyPair = ML_DSA_87.generateKeyPair()
+        const keyPair = signingUtil.generateKeyPair()
         return new MLDSAKeyPairIdentity(keyPair.publicKey, keyPair.privateKey)
     }